pyproject.toml
[project]
name = "ldr2svg"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "pillow>=12.1.1",
    "svgwrite>=1.4.3",
]

[project.scripts]
ldr2svg = "ldr2svg.ldr2png_svg:main"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.15.4",
    "ty>=0.0.19",
]

[tool.pytest.ini_options]
pythonpath = ["."]
markers = ["slow: requires OpenSCAD (deselect with -m 'not slow')"]
